// src/types/pressets/pressets.type.ts
import { ThemePagination } from "../theme/theme.interface";

export type PresetsType =
    | "white"
    | "lightBlue"
    | "blue"
    | "dark"
    | "roundedRich"
    | "roundedSpace"
    | "roundedAuroraNebula"
    | "roundedDeepSpaceVoid"
    | "roundedVoid"
    | "roundedBlackHole"
    | "roundedSolarFlare"
    | "roundedInferno"
    | "roundedTrimstone"
    | "roundedAbyssal"
    | "roundedOceanDepths"
    | "squaredForestMoss"
    | "squaredCyberpunkNeon"
    | ThemePagination;